Mostly fixed sizing & positioning.  Added H-button
The sizing and positioning of most of the elements on the page has been fixed.
Still working on fixing the sidebar to properly fill the viewport.

A homepage button has been made using CSS (didn't feel like making a image for it instead, plus it was good practice for manipulating CSS properties).
For now, it links to the main branch's homepage ("index.html").