Re-added images, select menu, and updated sidebar
Images have been re-introduced to the html file, alongside new stylings in the css file.
Stylings have also been re-organized based on how they appear in the html file.

The sidebar now links to the various flag categories (sorted by initial letter of the country).
A "back to top" button has also been added to the sidebar.

Currently the dropdown menu that should link to a specific country's flag is both non-functional and is the wrong location.